// src/app/features/destination/models/destination.model.ts
export type Destination = {
  id: string;
  country: string;
  continent: string;
  cityStart: string;
  cityEnd: string;
  startAt: Date;
  startEnd: Date;
  duration: number;
};

export type DestinationDTO = {
  id: string;
  country: string;
  continent: string;
  cityStart: string;
  cityEnd: string;
  startAt: Date;
  startEnd: Date;
  duration: number;
};